+++ packages/react/src/components/Filter/index.ts
@@ -1,2 +1,6 @@
 export { default } from './Filter';
-export type { FilterProps, FilterItem, FilterType } from './Filter.types';
+export type {
+  FilterProps,
+  FilterItem,
+  FilterType,
+} from './Filter.types';